+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# UNRELEASED
+
+- Updating postal code input to consider strings of length 2 to be valid input to support UK postal code format `XX XXX`.
+
 # 10.0.2
 
 - Update `credit-card-type` to 10.0.2